Rename breakout-room folder to ocp-2023 for ocp2023 template
